Add support for handling entire libraries of classes
There is an important TODO near the Collection definition in ast.py.
Basically I am not sure if the current implementation is as good as it can
be, especially because the Modelica specification seems to hint in a
particular implementation using nested classes. We however are using a
collection, which can contain files, which in turn contain classes. See
also commit 0057cb5d7fee55b96f845a114db69a19fa899286 , which we probably
cannot handle at the moment for this exact reason.