Fix compilation in model caching on Windows
Instead of relying on environment variables and calling the compiler
directly ourselves using sys.call(), we use the distutils ccompiler
package to do it for us. This will automatically locate the right
(supported) compiler.

We might have to detect what compiler is being used to set the right
compiler flags. For now it works with GCC on Linux and MSVC on Windows.