Use pickle directly instead of shelve
Shelve is annoying in that it is not cross-platform per se (which would
make the os.name check fail), and that the filenames are not easily
controlled either. The default fallback on Windows (dbm.dumb) for example
makes 3 files (.bak, .dir, .dat) whereas on Linux one commonly gets one
file without any extension.