Use multi-staged flattening approach
This commit reworks the flattening into multiple stages. The lookup
process is now also tree based, as that fits the initial stage of
"instantiation" well.

There is some new code for handling of redeclarations and modifications,
most of which is found in the build_instance_tree() routine. The other two
important routines for flattening (flatten_extends() and
flatten_symbols()) are essentially unchanged in their logic, except for
doing most editing of symbols/classes in-place.