Fix scope tracking of modifications
Before we checked whether a component reference existed in the current
class's symbol list. This caused the right hand side of modifications like
"foo(bar=bar)" to be flattened as well. So instead of ending up with
"foo.bar = bar" we ended up with "foo.bar = foo.bar".

This commit changes the behavior of the flattening in the following ways:
1. Modifications are not applied before symbols are flattened, but instead
   during this stage.
2. The component reference flattener (part of the symbol flattening call)
   skips processing of modifications.
3. We apply any modifications after symbol flattening, if the scope of the
   modification matches that of the current class. We do this by checking
   the full reference of both against each other.

Note that the current way of shifting modifications to child components
(classes or symbols) is the way it is outlined in the Modelica
specification (version 3.4, section 5.6.1). There is something to be said
for trying not to shift modifications (with scope), but it might be
possible we will not be able to comply with the specifications in other
aspects in the future.

Closes #38