Remove stored reference to root
Storing a reference to root complicates the (proper) implementation of
copy_including_children(). To avoid copying the _entire_ tree, we would
have to disable the root reference on all children as well. Either that,
or we would have to override the __deepcopy__ behavior, which then only
works when deepcopying but e.g. not when pickling.