Fix insane memory usage due to scope
Because we do not apply all modifications before flattening symbols
anymore, and instead keep a reference to a whole bunch of instanceclasses
around, the deepcopy() statements in the code would end up copying the
entire tree many times.

As we currently only compare on full reference of the scope (and not the
id() of the instance) when applying symbol modifications, we can get away
with skipping the copying of the scope by implementing our own deepcopy.