Fix constant references in Python 3.5
We were unknowingly assuming that we would enter the "classes" attribute
before the "symbols" and "equations" attributes. Because looping over
attributes in Python 3.6 will loop over them in declared order, this bug
did not show up. In Python 3.5, the order of looping over attributes is
not deterministic, which resulted in test_constant_references sometimes
failing, and sometimes succeeding.