Fix badges to show status of master
If we do not explicitly track the master branch, the badges will show the
status of the latest build, which could be e.g. a pull/merge request.

The Travis and AppVeyor badges are also changed to shields.io versions,
such that we can label them differently (instead of both being "build
status").