Group equations and statements in if/else branches
Instead of figuring out in the backends what conditions belong to what
equations/statements, and whether or not an else-branch was present, we
now group the contents of a branch in a block in the grammar/parser.

Furthermore, in the case an else branch was present, we add a condition
equal to "True", such that the number of conditions is always equal to the
number of blocks.

Note that the Modelica specification has certain requirements about
if/else constructs. For example, about the number of equations being equal
in each branch, or about whether an else branch is required.