Fix expansion of MX equations
According to Joris Gillis (CasADi co-author), ca.matrix_expand() is
broken and will not be fixed. Instead, we will then rely on regular
.expand() functionality to turn MX into SX, without the "graceful"
fallback that ca.matrix_expand() was supposed to offer. From initial
tests, it seems that most/all commonly used functionality is supported
by .expand().

Note that support for CasADi 3.4.0 is needed, as 3.3.0 and earlier do not
allow empty SXFunction (i.e. a function with no outputs).