More aggressive alias checking if not expanding MX
If expand_mx is False, the expansion of vectors results in complicated
expressions which are hard to analyse for aliases. In that case, we expand
the (scalar) equation to SX to see if it describes an alias.

Also note that we still do fast checks first, i.e. those operating
directly on the MX graph. The slowest operation (ca.substitute) is only
used as a last resort. This way the slowdown to the alias detection is
minimal (1.0 - 1.5x), even for very large models with 100,000+ alias
equations.

These extra checks typically result in 1-3% more aliases being detected
when expand_mx is False. Generally it results in the same number of
aliases being found as when both expand_mx and expand_vectors are True.

With these more aggressive checks, the test on "DelayForLoop" now no
longer differs in behavior with the setting "expand_mx". The two tests
have therefore been merged to reduce duplicate code.

Closes #55