Fix variable metadata function for MX expressions
The variable metadata function is used to store and load variable
metadata. For a size and speed improvement it can turn an expression
into its affine form. The check whether this would be possible was
however incomplete, resulting in a cached model not always having the
same variable metadata as a freshly parsed one.

This typically occurred when alias detection was on, as alias detection
can produce fmin/fmax operations for the min/max attributes. Those
fmin/fmax operations were lost in making the expression affine.