Fix nested array indexing
Indexing an element nested2[1].x, with x itself an array, did not work
properly. No errors were raised, but the result was not equivalent to
using nested2[1].x[:].

To fix this, we now keep track of the dimensions of both nested2 and x
separately, and also how they are indexed. In the above outlined
example, we detect that x is an unindexed array, and we add the slicing
operator [:] ourselves.

Keeping track of dimensions also enables us to detect wrongly indexed
variables, e.g. a user trying to index a scalar, or not specifiying the
right number of dimensions.