Improve handling of modelica delay() operator
We now reconstruct delay_arguments in load_model(), and handle delay arguments
in simplify(). This makes the api consistant, regardless of if the model
is cached or not.

We also now support the delay operator being used within for loops, as
well as constant inputs in the delay duration expression.