Expand vectors first when also expanding MX
When we expand MX to SX (with expand_mx=True), we want to expand as soon
as possible. The expansion of vectors turns into a complex MX graph,
which makes debugging and alias detection difficult.

We get rid of the complex graph by expanding to SX. However, further
simplification steps and our API rely on symbols being MX. We therefore
turn the SX graph back into an MX graph by rebuilding it from scratch.

Note that the expand_vectors code is not changed, just moved to a
separate function.