Fix expand vectors symbolic attribute error
This fixes a regression introduced in 901af4cc.

If a state had a shape that was not (1,) and symbolic parameters in its
attribute expressions, we would attempt to cast the sybolic expression
into a numpy array, throwing an error. We now check for symbolics in the
attribute expressions as well before casting to numpy array.

Previously, the tests did not cover this case, which is why the
regression was able to sneak in. This commit also adds a test for this
case.