Do not handle annotations in backend
Annotations cannot and should not be handled in the CasADi backend. We
therefore use a custom TreeWalker that skips this node. The approach of
skipping certain children like this is made more accessible by defining a
"skip_child" method.

A test case is added that would fail in the CasADi backend when it tried
to handle an expression inside an annotation.