Check symbol modification target existence
It could happen that a user would apply a symbol modification to e.g.
"my_var", whereas that should have been "My_var". This would silently
fail without applying any modifications at all, which can be very
confusing. Now we check that the targets of the to-be-applied
modifications actually exist.