Optimize getting aliases in AliasRelation
It is about 10-30x faster to prevent the instantiation of an OrderedSet.
If an element is in the aliases dict, the latency would be about 1.5 us
before versus 75 ns now. If an element is not in the aliases dict, the
additional delay caused by the explicit check is negligible.

Note that although these numbers seem small and negligble at first, they
can add up quickly for models with 100,000+ variables.