Optimize reading variable metadata from cache
For large models we would have very large variable metadata matrices,
with 100,000+ rows. Much of the metadata will be independent, which
means that a lot of accesses will be performed to what used to be a
bunch of CasADi DM matrices. This is much slower than indexing a NumPy
array, even with the added cost of converting it once.

For a model with about 100,000 variables, the transfer_model call went
from ~4 s to roughly ~2 s.