Fix bug where flattening takes exponential time
In some cases, a reference to the root of the ast tree would remain when
making a deep copy. Each symbol instantion would then take exponentially
more time (essentially doubling with every copy_including_children call).

A typical case is when a symbol has a type that (when looked up) is an
InstanceClass contained in the class we instatiate the symbols in.

To solve the issue, we make sure that all symbol types do not have any
references to their parents when making the copy, and putting everything
back together when the copy is done.

Closes #141