Temporarily disable CI for MacOS
The tests on MacOS fail because the latest version it installs is now
Python 3.7, which is not fully supported yet by all our dependencies.

It has also proven quite difficult to get MacOS testing to run reliably
without ways of testing in e.g. a VM. So we disable CI for MacOS until
that option becomes available, or someone with access to MacOS steps up to
make it work again.