Fix for loops with delay and non-indexed variable
If a non-indexed (scalar) variable was used inside the for loop where also
an delay() operator was used, the handling of the delay expression would
fail. Note that the scalar variable did not need to be used in the same
equation as the delay() operator.

The relevant test cases have been adjusted such that they would have
caught this bug.