Fix delay expressions for loops not starting at 1
We only make the mapped delay symbol as large as it should be. In case the
loop is over "i in 2:3", then the delay symbol would get a size 2. The
variable "s.indices" is still [1 2] however (0-based), meaning that would
get an indexing error. To fix it, we just slice the entire delay symbol.

The relevant test cases have been amended such that they would have caught
this bug.