Fix unwanted deepcopies of entire tree
The parent attribute of a class is only there to be able to walk the
tree. One cannot apply modifications to a Class's parent, so excluding
it from the deepcopy is a safe way to avoid unecessarily copying the
entire tree.

The "copy_including_children" method was there to take care of this, but
it turns out that lingering references remain. See e.g. commit 5a743f30
which tried to fix this for the "type" attribute of InstanceClass
objects.