Add insertion of array indices at correct index
The indices for nested array variables when expanding vectors were
always inserted at the end (c.b.x[1,2,1] etc.). It is more intuitive if
the naming of the variables is consistent with modelica (c[1,2].b.x[1]
if c was a 2d array and x a 1d array). The indices are now inserted
directly after the array variable(s).