Fix indexing of nested arrays in equations
If you have nested arrays like a[3].x[3] in equations, then a[1].x =
{1,2,3} should be equivalent to a[1].x[:] = {1,2,3}. The result should
be a[1].x[1] = 1, a[1].x[2] = 2 etc.  This was not the case before.
Pymoca now stores the indices of symbols in equations separately for
each nested element, allowing it to detect cases where the [:] was
omitted and quietly inserting it when necessary.