Fix better indexing for nested equations
When defining a model with nested symbols such as
b.a.x, an equation can be defined in one of the
lower level classes, for example a.x = 2. When
defining such equations for nested arrays the new
indexing did not yet function properly.

Note: dimensions higher than 2 are not yet
supported for equation symbols.