Fix dimension list depth
The dimensions of a Symbol should be a list of lists, to record the
dimension of each symbol when flattening nested symbols. Previously,
the dimension of a Symbol / ComponentRef was a simple list, and turned
into a list of lists during the flattening. This could result in a list
of list of list etc. Now the dimension is already a list of lists when
creating the Symbol in the parser, which can simply be combined during
flattening.

Closes #155