Coerce values to Python type when expanding
When not expanding, many symbols are already their Python type if they
can be converted to it. When expanding however, we should take care to
do the same. For example, a ca.DM or ca.MX matrix with a size larger
than one could easily allow us to convert its values to float as well.