Fix infinite loop in replace_parameter_expressions
In the replace_parameter_expressions simplification step, it was possible
to end up in an endless loop when there are a lot of parameters that have
symbolic values. With the expansions of vectors first (when expand_mx is
True) introduced with 028093f, it is more likely to end up with such
cases.

The CasADi ca.is_equal call will terminate when its maximum depth is
reached, and return False. The limit of 100 introduced in adbd468e is not
enough in practice to compare two MX graphs. We therefore increase this
number significantly.

A safeguard is also introduced to prevent an infinite loop from happening
again, by raising an Exception after (currently) 100 iterations. This is
aimed to catch situations (e.g. CasADi bugs) where for some reason x != x.