Fix initialization of derivative symbol in some cases
If the first encountered expression of a certain derivative was something
like "der(x[1])", the symbol for the derivative was not properly
inititialized.

Note that if the first encountered expression was instead something like
"der(x)" or "der(x[i])" (like in for loops), the initialization was done
correctly. Encountering an expression like "der(x[1])" afterwards would
then also not lead to any problems.

Closes #163