CasADi generator:  Add _pymoca_interp{1,2}d builtins
These are useful helpers to do quick linear interpolation.
The _pymoca prefix is there to indicate that these functions
are not part of the Modelica specifications.

We use the order_keys() functionality added in the previous commit
(70e37417), because the CasADi interpolation primitives require the
numerical values of the data points to be known at creation time.
Hence, these must be fetched from the symbol definitions. This, in turn,
requires that the symbol definitions are available before the equations
are parsed.