Fix redeclarations with classes inside flattening class
Compare the RedeclarationScope.mo to RedeclarationScopeAlternative.mo.
They should both give the same results, but the flattening of the latter
did not work correctly before this commit. The reason was that the
classes HQBase/HQZ/Channel inside ChannelZ (the class we flatten) are
handled before "Channel c" symbol declaration. The types of the symbols
inside these classes was therefore already resolved to elements in the
InstanceTree. Any redeclarations did therefore no longer do anything on
the symbols.

The fix is to make sure we also update symbol type references to the
redeclared class.