CasADi: Improve import latency for cached models
There are a few imports in the CasADi backend API that take in the order
of 100 ms. One of which is the (indirect) import of the ANTLR generated
ModelicaParser/Lexer files, which deserialize a binary blob upon import.
We can avoid this import for cached models by lazily importing the
`parser` module.

The other high-latency import is the distutils compiler import, which
will locate the system's copmiler upon import. This import is only
necessary when using codegen to compile a model, so we can instead
lazily import it then.