Tree: Give information about failing connect equations
Before we would only raise an AssertionError when we were not able to
resolve a ComponentRef in a connect clause. This error would not provide
any information about _what_ connect statement was wrong, which made it
hard to debug. We now print out some (not perfect) information about the
failing statement.