CasADi: Coerce attribute type to variable type
If, for example, the variable is of type Integer, we now try to coerce
the attributes (start, min, max, value, nominal) to int. Integers do not
support infinity or NaN, so we still have to use float in those cases.

Note that we only do this for scalar integers and floats for now.
Booleans are not yet handled, nor are arrays.