CasADi: Fix small nominals on aliases by making default nominal zero
Before this commit, the default nominal was 1.0. If a user explicitly
set a nominal of e.g. 0.1 on a variable, but would not set a nominal on
an alias thereof, the alias detection would end up with a nominal of
1.0.

This problem is easily fixed by setting the default nominal to 0.0,
which would mean that it is not set.