CasADi: Fix expand vectors for fill()/eye()/...
A fill/eye call would result in a ca.DM object, and not a list as was
asserted in model.py. The expansion code is fixed to account for ca.DM
and np.ndarray types as well. In line with 7f9fd64 we also force a
conversion to the Variable's Python type (float or int).