CasADi: Warn if start attribute of alias conflicts
Instead of just blindly overwriting the start attribute with that of the
lastest alias, we should check if there is a conflict to avoid suprises.

Note that we cannot exhaustively check for conflicts, because if the
user sets the canonical state's start attribute explicitly to zero, it
will still overwrite it with that of the alias.

Also note that the order in which we loop over aliases is random (due to
the usage of a set), so the order in which messages appear and what
conflict they report can be different. This also means that, in the case
the user does not set a non-zero start attribute on the canonical state
(that is the first encountered alias), but does do so on multiple of its
aliases, the eventual start attribute of the canonical state cannot be
known in advance. This situation should be prevented by making sure that
all aliases have matching (or not-set) start attributes. The warnings we
added in this commit will help with that.

Closes #208