SymPy: Fix defaulting to start when value is set to 0.0
A statement like `if value` will do a falsy check on `value`. This means
that a value of `0` or `0.0` will also return False, resulting in the
start attribute being used (which may be non-zero). This holds for
Python, and also seems to hold for Jinja. So we have to make sure to
explicitly check for None (or "none" for Jinja).