AST: Represent unspecified start attribute with None
It is very useful to know whether the user has a specified a start
attribute or not. By forcing the start attribute in the AST to 0, we
lost this information, and could therefore not do proper merging of
alias start attributes in the CasADi backend.

It is up to the various backends to make sure that the start value, if
not set by the user, is given the default value of 0.0. Note that this
is just like the way nominal, min and max were handled already.