CasADi: Allow/fix multiple calls to simplify
In the case that the three options `eliminate_constant_assignments`,
`replace_constant_values` and `detect_aliases` are set, multiple calls
to simplify would fail. The simplify method was not intended to be
called multiple times, but there are advantages to being able to do so,
especially with the combination of these three options.

The reason for the failure with the three options mentioned above is
that in `replace_constant_values` we would remove Variable instances
from the list of constants. When allowing multiple simplify() calls we
should then also make sure that `self.alias_relation` is cleaned up as
well, and no longer contains any references to the removed constant
Variable or any of its aliases. Lingering references to non-existing
variables are not desired (if not plain wrong), and would also lead to
processing in the `detect_aliases` block failing.

Closes #214