CasADi: Support iterative simplification option
The compiler/simplification option `iterative_simplification` is added
to repeatedly simplify a model until convergence. Multiple iterations of
simplification are needed in some cases for convergence, especially when
the removal/detection of aliases will cause additional constants to be
able to be eliminated (which in turn will allow for new aliases to be
detected, etc).