CasADi: Fix aliasing of algebraic states to constants
The aliasing of algebraic states to parameters already worked as
expected, but that for constants did not yet. This commit fixes the
latter such that aliasing behavior to constants is similar to that of
parameters.

Two tests have been added. Note that only ConstantAlias model failed
before this commit, and that the ParameterAlias model is added for
completion and easy comparison (i.e. expected aliasing behavior is very
similar in both tests).

Closes #218