CasADi: Support string parameters and constants in model
String parameters and constants were not supported yet, but can be
useful e.g. for identifying the type of component after having flattened
a model. Note that string parameters/constants cannot appear in any
CasADi expressions/equations/assignments. Three options were identified
in which these variables could be included in the model:

1. `model.parameters/constants` and contains all parameters/constants
regardless of type. CasADi functions are a function of all
parameters/constants.

2. `model.parameters/constants` contains all parameters/constants
regardless of type. CasADi functions are a function of only the numeric
parameters/constants.

3. Add `model.string_parameters/constants` as a separate
parameter/constant group for non-numeric parameters. Not an input to any
of the CasADi functions.

This commit implements option 3. The reason this option was chosen over
the other two is that it is somewhat confusing to have a CasADi DAE
function that is a function of `[states, alg_states, parameters, ...]`,
but then somehow always have it be the case that it is only a function
of a subset of those. Furthermore, if it is a function of all
parameters/constants (option 1), we would force users to provide some
nonsense numeric value for these string parameters/constants. If it is
explicitly only a function of a subset of parameters/constants like
`[states, alg_states, *[p for p in parameters if isinstance(p,
Variable), ...]` (option 2), the calling signature is not entirely
obvious for users and applications using Pymoca.

Do note that the the Pymoca API is still unstable (and will be that way
for now). That means that future insights/experiences could still lead
to a different implementation of string parameter/constant support.