CasADi/Simplify: Add option to (dis)allow aliasing derivative states
Some software handles derivate states rather differently from regular
states. For example, the derivative states might not actually be
individual symbols collocated in time, but instead just the difference
between two collocated instances of the associated state (divided by a
time interval).

When an algebraic state is an alias of a derivative state, the canonical
state is the derivative. The processing of the aliased (algebraic) state
can then easily fail when it relies on there being associated collocated
symbols. This commit adds an options to disallow aliasing to derivative
states for such software.