gen_casadi_test: Fix instantiating np.array with list of MX objects
Starting from, NumPy 1.20 behaves differently when instantiating arrays
from array-like objects. NumPy used to be agnostic about the type of the
object inside the list (MX in our case), but now it actually tries to
interpret it as an array if it can. In this particular case, we
deliberately _did not_ want this behavior. We use something akin to the
work-around described in [0] to get the old behavior back.

For more details, see:

[0] https://numpy.org/devdocs/release/1.20.0-notes.html under "ArrayLike
  objects which do not define __len__ and __getitem__" or

[1] https://github.com/numpy/numpy/pull/17973