CasADi/Simplify: Cast to Python type when substituting metadata
When substituting metadata, we prefer to end up with the type of the
corresponding Variable.

For instance, when the expression `0.25 * x ** 2` is resolved for `x =
1.0`, we would end up with a ca.MX value instead of a float. These MX
values cannot directly be used in NumPy array expressions (both before
and after NumPy 1.20). This commit makes sure that we end up with the
correct Python type.

Note that we skip casting for booleans, as their min/max/nominals are
rarely used. For integer variables we make sure that NaN and Inf are
converted to Python floats, because the Python int type does not support
these special values.

Also note that we cannot convert directly from ca.MX to int, we have to
go via float.