Tools/compiler: Add options
This is a start at a more conventional Modelica compiler command line
tool for easier use of pymoca without Python scripting. It's also a
basis for upcoming improvements in Modelica library parsing, providing a
tool to parse entire libraries and print elapsed wall clock time and
a count of parse errors and messages with 2 levels of verbosity.

The compiler tool can be called from the command line as before and now
also can be imported as is done for the test cases. Options have been
changed and are incompatible with the previous version.

It currently supports casadi and sympy backends, but is intended to be a
start at supporting additional backends. A good improvement would be to
implement a pluggable backend API to be used by the compiler tool. This
would need some refactoring of the casadi api. Also the xml backend
could be easily added if it is in use.

All messages are printed to stderr to be in the same stream as those
from the ANTLR parser. Really only errors should go there, but I think
is better for now until we can figure out how to get logging to send to
multiple streams.

Some other changes included in this commit:

* Add test cases for the compiler tool which include adding a Package
  directory under test/models and a copy of Spring.mo to support testing
  of the casadi code in the compiler tool.

* Change the sympy backend to take the options passed through the
  compiler tool. They are currently unused by sympy but the options
  are envisioned as part of a new general API for backends (TBD).

* Fix type annotations flagged by mypy while changing the sympy
  backend.

* Add files to .gitignore: *.pymoca_cache, pytest.ini