Modelica: Add Modelica Standard Library submodules
This adds submodules to track maintenance branches of both MSL 3.2.3 and
MSL 4.0.x to use for testing. Pymoca started in the MSL 3 days and is
probably most compatible with that version. However, MSL 4 is broken up
into smaller files so it is easier to pick targeted test cases that
avoid features that Pymoca doesn't currently support. It also is the
future of Modelica. I believe many libraries are still not compabatible
with MSL 4 and the latest language spec, so I expect to need to test
against both versions and is the reason for including both here.