Modelica: Limit CasADi symbol attributes
Expanding coverage of the Modelica language is adding attributes like
`ast.Symbol.quantity` used in units from the Modelica Standard Library.
To future-proof the CasADi code, let's specify the attributes that the
CasADi model supports instead of keying on the attributes we don't
support (which may expand in the future).

See PR #36 and d8f7bea where the previous code was merged.