Bump minimum python to 3.7
Github Action no longer supports 3.5, and it has also been end-of-life
for over 2 years at time of writing. We bump the minimum version to
Python 3.7 (the lowest version still supported).

We also get rid of the manual version check in setup.py, which no longer
works for Python 3.10 and higher (due to "10" < "7"). We just rely on
"python_requires" from this point, which should have wide enough
adoption.

Also note that we are not bumping the maximum supported versions, as
the latest Python version supported by CasADi is still only 3.9.