Tree: Fix scoping of inherited symbols
This fixes cases where inherited symbol type lookup fails because the
scope of the type was lost during instantiation.  One example where it
fails is in flattening MSL 4 `Modelica.Electrical.Analog.Basic.Resistor`
where the symbol type name `PositivePin` is referenced in an inheritance
hierarchy two levels deep.

See Modelica Language Spec 3.5 section 5.6.1.2 and related sections.

This implementation adds a parent attribute to the `ast.Symbol` class
that is set during creation of the AST. If local lookup fails, the
original symbol parent is then used as the starting point for symbol
type lookup in `tree.build_instance_tree()`.

A test case is added that is similar to MSL 4 Resistor that failed.