SymPy: Fix generator duplicate variable creation
Fixes a problem where duplicate SymPy symbols are created in the
generator for Modelica output variables. SymPy 1.12 aparently added a
check for duplicate symbols. SymPy throws an exception
`ValueError: duplicate symbols given` in `sympy.solve()` that causes
`test_estimator` in `gen_sympy_test` to fail.

Since outputs are either variables or states, the SymPy symbol already
created for the variable or state is now used for the output.