CI/CD: Add GitHub release action
This adds a GitHub release action to track along with the PyPI releases.

It uses `softprops/action-gh-release` which is 1 of 4 actions
recommended by the now archived GitHub action
[`actions/create-release`](https://github.com/actions/create-release)
and has the most stars and contributors of the 4.

GitHub auto-generated release notes is turned on with no other options
to keep maintenance to a minimum while Pymoca is at its current
level of development. I expect we will want to have some human
intervention in creating release notes as we get closer to the 1.0
milestone and it looks very doable to gradually add more of that.