style/flake8: Fix all but E501 line-too-long errors
We fix (almost) all types of errors that flake8 checks for. We exclude a
few checks, see also setup.cfg:

- E203: Space before colon (not PEP-8 compliant, and conflicts with
  black)

- W503: Starting lines with operators. We prefer it this way actually,
  and black also likes it like this.

We also _temporarily_ ignore line-length errors, as many of those are
related to issues that need to be resolved first like typing annotations
(currently done as comments).

Lastly, there are a few inline `noqa` exceptions of the types:

- F841: unused variable. Sometimes very useful for debugging, so mostly
  present in tests

- F401: unused import. Mostly related to typing annotations, which
  should be resolved in a future commit that moves those from comment to
  Python-3 style annotations.

- F741: Variable "I" used, only in a test to match the naming of
  Modelica symbols. We should avoid using these types of names in
  Modelica models, as to avoid such flake8 errors in the corresponding
  Python code.

Note that we get rid of the complexity checks, as that would just lead
to a whole bunch of ignores. And then also get rid of the `noqa:
too-complex` in the two files that were _not_ excluded from flake8 tests
before this commit. Eventually it might make sense to re-introduce this
complexity test, but that is a less trivial reformatting/refactoring of
code compared to this commit.

Also note that black reformatted a single piece of code in
`sim_scipy.py` due to the changes related to flake8.