style: Prepare for black, keep arguments over multiple lines
There are a few places in the code where we split arguments over
multiple lines. To make sure that black formats the code in a similar
fashion, we have to append a comma to the last line/argument.

Note that we do this in a separate preparation commit, because we want
the overall black-formatting commit to be just that; the changes
made by black.