CasADi/Simplify: Fix `resolve_parameter_values` for list/array values
The symbol "value" was not instantiated when the parameter value was a
list. This was not caught before, because there was no test for such
cases. This commit adds such a test case next to fixing the bug.

Furthermore, we were only looking at lists and not also at NumPy arrays.
We force convert from list to array in this if-branch, and also in some
places in the code. We therefore have to make sure the condition catches
both lists and np.ndarrays.