parser: Add caching of parsed Modelica files
This commit introduces a caching mechanism for parsing of Modelica
files. The cache is stored in a SQLite (system-wide) database file and
cache entries are pruned based on their last access time (30 days by
default). We purely cache the tree of a single Modelica file, not a
merged (library) tree or flattened model.

Because the Pymoca version is part of the uniquely identifying key for
each entry, entries that have not been accessed for a certain number of
days are removed from the cache to avoid an ever-growing database.

Additionally, if a work tree is dirty, caching is disabled. This is to
ensure that changes made to the code are always reflected in the parsed
output, even if a cached version of the code exists.

As far as performance improvements due to this commit is concerned:
parsing is a model is generally sped up by a factor of 4 - 200. On the
faster end if there are a lot of comments / annotations, and on the
smaller end for small files. Large files with very few comments can
generally expect a 10x speed up in the parsing step. Note that other
factors like building/merging the tree (e.g. `tree.extend()`) and
flattening can become the limiting factor for compiling a model.