Remove generated test files from repository
These files are generated when running the tests. Typically, their style
does not adhere to black's standards, and thus this leads to seeing them
in the modified section when running `git status`, and potentially
accidentally committing them. As there is no need for them to be
version controlled, we remove them.

Note that we need to keep the folder around for the tests to succeedi
(they do not make it if it does not exist), so we use a .gitkeep file in
there.

Closes #291