Style: Use pre-commit for black and flake8
This uses the pre-commit Python package to run black and flake8 as git
pre-commit hooks: https://pypi.org/project/pre-commit/.

This allows the flake8 and black style checks to be done locally to
catch issues before pushing to the remote repo. You will have to fix
them before the local commit can proceed if you have installed
pre-commit.

Instead of running black or flake8 separately, use the pre-commit CLI
tool `pre-commit run` to run them so you get the same config as is being
used in the GitHub action.

Since we are using pre-commit to run style checks locally with a git
pre-commit hook and remotely with GitHub actions, we no longer need
to include it in the tox config. Use `pre-commit run` if you wish
to manually run the style checks, but the git pre-commit hook will
run them automatically, so it is generally not necessary to do it
manually. Also, IDEs have tools for automatically doing style and lint
checks and sometimes fixes.

There are instructions for using black and flake8 with pre-commit in
their docs:
https://black.readthedocs.io/en/stable/integrations/source_version_control.html
https://flake8.pycqa.org/en/latest/user/using-hooks.html

Note that `exclude` was changed to `force-exclude` in the black section
of `pyproject.toml` per the instructions from the black document above.
This allows it to play nicely with pre-commit hooks.

Also note that flake8 does not have a force-exclude option like black,
so we need to feed it through the pre-commit config file in addition
to specifying in flake8's config in setup.cfg.