Update README
Update the README for the current state of the project and give a
little more overview and updated plans.

Remove the stale conda instructions and go with just pip.

Remove link to the broken xml notebook example. The ModelicaXML
GitHub repo has not be touched in 9 years, and Google shows
no recent activity on the subject, so it appears to be a dead end.