Use setuptools_scm instead of versioneer
We are already using setuptools for many things, and setuptools_scm
seems to become the defacto standard for automatic versioning.

Versioneer suffixed with `.dirty` when there were local changes, and
that `setuptools_scm` uses a suffix with the date in it by default, e.g.
`.d20240124` and increments the preceding version number to the next
guess. We therefore modified the checks on "dirty state" for caching
purposes accordingly.

Note that it is hypothetically possible to make setuptools_scm behave
like versioneer, and suffix with `.dirty`. This was however not easy to
get to work correctly, and seemed to fail in some cases depending on how
one requests the version.