Use hatchling and hatch-vcs instead of setuptools + versioneer
We were looking at ways to remove versioneer, and switch to
setuptools_scm which integrates more nicely with modern Python
(pyproject.toml and the like). The issue with setuptools_scm is that is
difficult to get it to work reliably to detect "dirty" local
repositories in editable installs. This is something we need to
invalidate (or skip) caches in dev environments.

Hatchling is a build system, part of the package manager 'hatch', and
has become a popular alternative to setuptools being a lot more
lightweight and focused (again) on modern Python approaches. There is a
plugin hatch-vcs which uses setuptools_scm under the hood to detect
versions. There is also a "footgun" example for 'hatch-vcs' to provide
us the necessary "dirty" detection we want in editable installs,
matching Versioneer. The caveats mentinoed were probably applicable to
versioneer as well:

> This hybrid approach is somewhat of a footgun: it involves distinct
  version detection mechanisms between development and deployment.
  Ideally you should always remember to reinstall the package whenever
  checking out a new commit so that you can simply use the standard
  importlib.metadata.version mechanism. In constrast, the hybrid
  approach is unsupported, so it must be used at your own risk.

Note that Versioneer suffixed with `.dirty` when there were local
changes, and that `setuptools_scm` (and hatch-vcs by extensions) uses a
suffix with the date in it by default, e.g. `.d20240124`.

Also note that, by default, setuptools_scm increments the preceding
version number to the next guess. We prefer to just have "post1.devN"
suffixes instead, as the version number guessing is often wrong. This
latter behavior is set by using 'no-guess-dev'.